refactor into a single processing function - add lock with default ttl - more robust Metadata comparison
